backend/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tax-buddy-backend"
version = "0.1.0"
description = "Hybrid AI workflow for automated tax return filing in India"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "python-multipart>=0.0.9",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "sqlalchemy>=2.0.30",
  "psycopg[binary]>=3.2.1",
  "python-dotenv>=1.0.1",
  "pillow>=10.4.0",
  "numpy>=1.26.4",
  "opencv-python-headless>=4.10.0.84",
  "pdfplumber>=0.11.4",
  "reportlab>=4.2.2",
  "jinja2>=3.1.4",
]

[project.optional-dependencies]
ocr = ["pytesseract>=0.3.13", "paddleocr>=2.8.1"]
ml = ["transformers>=4.44.2", "torch>=2.4.0", "datasets>=3.0.0", "seqeval>=1.2.2"]
dev = ["pytest>=8.3.2", "httpx>=0.27.2", "ruff>=0.6.2"]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
